Database Code Update
After a doozie of a day working with the database we decided that the
CLS_ID should be a random large number. Michael suggested we implement
this number in PHP which is what I have done here. The acutal PHP code
uses rand(x,y) to select a 5-digit class ID (that's the limit of the
rand function). We also added some views and switched the ordering to
prevent dependency errors.