SQL Query Update
Okay the sqlQuery function I added to functionlib is going to clean the
code up so much. I haven't replaced all the code with it yet but
basically everything function that uses the database can be shortened by
about 20 lines with this method. It returns true if the command worked
and, an array of the values if it is a SELECT statement, and null if it
returns nothing/fails. This should make the code super simple.